requirements.txt
asgiref==3.7.2
blinker==1.6.2
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
Django==4.2.1
Flask==2.3.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
ndg-httpsclient==0.5.1
pusher==3.3.2
pyasn1==0.5.0
pycparser==2.21
PyNaCl==1.5.0
pyOpenSSL==23.1.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.2
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.3.0
gunicorn

python-dotenv~=1.0.0
beautifulsoup4~=4.12.2